Rename built-in compileShell to prevent conflicts with funfuzz
